Separate avatar to entity user
