+++ src/TTV/UserBundle/Entity/Avatar.php
@@ -1,0 +1,205 @@
+<?php
+
+namespace TTV\UserBundle\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\Validator\Constraints as Assert;
+
+/**
+ * Avatar
+ *
+ * @ORM\Table(name="avatar")
+ * @ORM\Entity(repositoryClass="TTV\UserBundle\Repository\AvatarRepository")
+ * @UniqueEntity(fields={"extension"}, message="Un avatar existe déjà avec ce nom")
+ */
+class Avatar
+{
+    /**
+     * @var int
+     *
+     * @ORM\Column(name="id", type="integer")
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="AUTO")
+     */
+    private $id;
+
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="extension", type="string", length=255, unique=true)
+     */
+    private $extension;
+
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="alt", type="string", length=255)
+     */
+    private $alt;
+
+    /**
+     * @var UploadedFile
+     * @Assert\Image(maxSize="200k", maxSizeMessage="La photo d'avatar ne doit pas dépasser 200k.")
+     */
+    private $file;
+
+    private $tempFilename;
+
+    /**
+     * Get id
+     *
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * Set extension
+     *
+     * @param string $extension
+     *
+     * @return Avatar
+     */
+    public function setExtension($extension)
+    {
+        $this->extension = $extension;
+
+        return $this;
+    }
+
+    /**
+     * Get extension
+     *
+     * @return string
+     */
+    public function getExtension()
+    {
+        return $this->extension;
+    }
+
+    /**
+     * Set alt
+     *
+     * @param string $alt
+     *
+     * @return Avatar
+     */
+    public function setAlt($alt)
+    {
+        $this->alt = $alt;
+
+        return $this;
+    }
+
+    /**
+     * Get alt
+     *
+     * @return string
+     */
+    public function getAlt()
+    {
+        return $this->alt;
+    }
+
+    /**
+     * Set file
+     *
+     * @param string $file
+     *
+     * @return Avatar
+     */
+    public function setFile(UploadedFile $file)
+    {
+        $this->file = $file;
+
+        if (null !== $this ->extension){
+            $this->tempFilename = $this->extension;
+
+            $this->extension = null;
+            $this->alt = null;
+        }
+    }
+
+    /**
+     * Get file
+     *
+     * @return string
+     */
+    public function getFile()
+    {
+        return $this->file;
+    }
+
+    public function getUploadDir()
+    {
+        return 'uploads/avatars';
+    }
+
+    protected function getUploadRootDir()
+    {
+        return __DIR__.'/../../../../web/'.$this->getUploadDir();
+    }
+
+
+    /**
+     * @ORM\PrePersist()
+     * @ORM\PreUpdate
+     */
+    public function preUpload()
+    {
+        if (null === $this ->file){
+            return;
+        }
+
+        $this->extension = $this->file->getClientOriginalName();
+        $this->alt = $this->file->getClientOriginalName();
+    }
+
+    /**
+     * @ORM\PostPersist()
+     * @ORM\PostUpdate()
+     */
+    public function upload()
+    {
+        if (null === $this->file){
+            return;
+        }
+
+        if (null !== $this->tempFilename){
+            $oldFile = $this->getUploadRootDir().'/'.$this->tempFilename;
+            if(file_exists($oldFile)){
+                unlink($oldFile);
+            }
+        }
+        $this->file->move($this->getUploadRootDir(), $this->extension);
+    }
+
+    /**
+     * @ORM\PreRemove()
+     */
+    public function preRemoveUpload()
+    {
+        $this->tempFilename = $this->getUploadRootDir().'/'.$this->extension;
+    }
+
+    /**
+     * @ORM\PostRemove()
+     */
+    public function removeUpload()
+    {
+        if (file_exists($this->tempFilename)){
+            unlink($this->tempFilename);
+        }
+    }
+
+    public function getWebPath()
+    {
+        return $this->getUploadDir().'/'.$this->getId().'.'.$this->getExtension();
+    }
+
+}
+
