+++ src/TTV/UserBundle/Entity/User.php
@@ -4,7 +4,6 @@ namespace TTV\UserBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Validator\Constraints as Assert;
 
@@ -56,26 +55,12 @@ class User implements UserInterface
     private $roles;
 
     /**
-     * @var string
-     *
-     * @ORM\Column(name="avatar", type="string")
+     * @ORM\OneToOne(targetEntity="TTV\UserBundle\Entity\Avatar", cascade={"persist", "remove"})
+     * @ORM\JoinColumn(nullable=false)
+     * @Assert\Valid()
      */
     private $avatar;
 
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="alt", type="string", length=255)
-     */
-    private $alt;
-
-    /**
-     * @var UploadedFile
-     * @Assert\Image(maxSize="200k", maxSizeMessage="La photo d'avatar ne doit pas dépasser 200k.")
-     */
-    private $file;
-
-    private $tempFilename;
 
     public function eraseCredentials()
     {
@@ -196,7 +181,7 @@ class User implements UserInterface
      *
      * @return User
      */
-    public function setAvatar($avatar)
+    public function setAvatar(Avatar $avatar)
     {
         $this->avatar = $avatar;
 
@@ -213,120 +198,4 @@ class User implements UserInterface
         return $this->avatar;
     }
 
-    /**
-     * Set alt
-     *
-     * @param string $alt
-     *
-     * @return User
-     */
-    public function setAlt($alt)
-    {
-        $this->alt = $alt;
-
-        return $this;
-    }
-
-    /**
-     * Get alt
-     *
-     * @return string
-     */
-    public function getAlt()
-    {
-        return $this->alt;
-    }
-
-
-    /**
-     * Set File
-     * $param UploadedFile $file
-     */
-    public function setFile(UploadedFile $file)
-    {
-        $this->file = $file;
-
-        if (null !== $this ->avatar){
-            $this->tempFilename = $this->avatar;
-
-            $this->avatar = null;
-            $this->alt = null;
-        }
-    }
-
-    /**
-     * Get File
-     * @return UploadedFile
-     */
-    public function getFile()
-    {
-        return $this->file;
-    }
-
-    public function getUploadDir()
-    {
-        return 'uploads/avatars';
-    }
-
-    protected function getUploadRootDir()
-    {
-        return __DIR__.'/../../../../web/'.$this->getUploadDir();
-    }
-
-
-    /**
-     * @ORM\PrePersist()
-     * @ORM\PreUpdate
-     */
-    public function preUpload()
-    {
-        if (null === $this ->file){
-            return;
-        }
-
-        $this->avatar = $this->file->getClientOriginalName();
-        $this->alt = $this->file->getClientOriginalName();
-    }
-
-    /**
-     * @ORM\PostPersist()
-     * @ORM\PostUpdate()
-     */
-    public function upload()
-    {
-        if (null === $this->file){
-            return;
-        }
-
-        if (null !== $this->tempFilename){
-            $oldFile = $this->getUploadRootDir().'/'.$this->tempFilename;
-            if(file_exists($oldFile)){
-               unlink($oldFile);
-            }
-        }
-        $this->file->move($this->getUploadRootDir(), $this->avatar);
-    }
-
-    /**
-     * @ORM\PreRemove()
-     */
-    public function preRemoveUpload()
-    {
-        $this->tempFilename = $this->getUploadRootDir().'/'.$this->avatar;
-    }
-
-    /**
-     * @ORM\PostRemove()
-     */
-    public function removeUpload()
-    {
-        if (file_exists($this->tempFilename)){
-            unlink($this->tempFilename);
-        }
-    }
-
-    public function getWebPath()
-    {
-        return $this->getUploadDir().'/'.$this->getId().'.'.$this->getAvatar();
-    }
 }
