+++ src/TTV/UserBundle/Repository/AvatarRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TTV\UserBundle\Repository;
+
+/**
+ * AvatarRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class AvatarRepository extends \Doctrine\ORM\EntityRepository
+{
+}
